refactor: create handlers for differents events
